+++ android-emulator/migration-tcp.c
@@ -134,7 +134,7 @@ static void tcp_accept_incoming_migration(void *opaque)
 {
     struct sockaddr_in addr;
     socklen_t addrlen = sizeof(addr);
-    int s = (unsigned long)opaque;
+    uintptr_t s = (uintptr_t)opaque;
     QEMUFile *f;
     int c, ret;
 
@@ -201,7 +201,7 @@ int tcp_start_incoming_migration(const char *host_port)
         goto err;
 
     qemu_set_fd_handler2(s, NULL, tcp_accept_incoming_migration, NULL,
-                         (void *)(unsigned long)s);
+                         (void *)(uintptr_t)s);
 
     return 0;
